# perf(dynamite_runtime): improve the performance of collecting bytes