running the `large_get` test with a 5GB payload resulted in a 10% performance gain (21s -> 19s) and 44s -> 39s for a 10GB file.
Tested on my dev machine. Performance gain on low end devices will probably be higher.